requirements.txt
python-binance
pandas
numpy
plotly
requests
pandas_ta
streamlit
deep-translator
